Fix problem with signed short values, use int instead.

+++ src/test/java/com/monstarmike/tlmreader/datablock/StandardBlockTest.java
@@ -12,7 +12,7 @@ public class StandardBlockTest {
 
 	@Test
 	public void testReadRpmZeroNoRpmHeader() {
-		short testValue = (short) 0x0000;
+		int testValue = 0x0000;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x6).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, null);
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue), standardBlock.getRpm(), 0.001);
@@ -20,7 +20,7 @@ public class StandardBlockTest {
 
 	@Test
 	public void testReadRpm1111ZeroNoRpmHeader() {
-		short testValue = (short) 0x1111;
+		int testValue = 0x1111;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x6).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, null);
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue), standardBlock.getRpm(), 0.01);
@@ -28,7 +28,7 @@ public class StandardBlockTest {
 
 	@Test
 	public void testReadRpm00FFZeroNoRpmHeader() {
-		short testValue = (short) 0x00FF;
+		int testValue = 0x00FF;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x6).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, null);
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue), standardBlock.getRpm(), 0.02);
@@ -36,7 +36,7 @@ public class StandardBlockTest {
 
 	@Test
 	public void testReadRpmFF00ZeroNoRpmHeader() {
-		short testValue = (short) 0xFF00;
+		int testValue = 0xFF00;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x6).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, null);
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue), standardBlock.getRpm(), 0.001);
@@ -44,7 +44,7 @@ public class StandardBlockTest {
 
 	@Test
 	public void testReadRpmFFFFZeroNoRpmHeader() {
-		short testValue = (short) 0xFFFF;
+		int testValue = 0xFFFF;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x06).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, null);
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue), standardBlock.getRpm(), 0.001);
@@ -57,7 +57,7 @@ public class StandardBlockTest {
 		HeaderRpmBlock rpmHeaderMock = Mockito.mock(HeaderRpmBlock.class);
 		Mockito.when(rpmHeaderMock.getPoles()).thenReturn(poles);
 		Mockito.when(rpmHeaderMock.getRatioInHunderth()).thenReturn(ratioInHunderth);
-		short testValue = (short) 0xFFFF;
+		int testValue = 0xFFFF;
 		byte[] testBlock = new DataBlockBuilder(0, STD_DATABLOCK_MAKER).setValue(testValue, 0x06).get();
 		StandardBlock standardBlock = (StandardBlock) DataBlock.createDataBlock(testBlock, rpmHeaderMock );
 		Assert.assertEquals(reverseEngineeredRpmValueFunction(testValue) / poles / ratioInHunderth * 100, standardBlock.getRpm(), 0.001);
@@ -87,7 +87,7 @@ public class StandardBlockTest {
 		Assert.assertEquals((87 - 32) / 1.8,  standardBlock.getTemperatureInGradCelsius(), 0.001);
 	}
 	
-	private double reverseEngineeredRpmValueFunction(final short rawShortValue) {
+	private double reverseEngineeredRpmValueFunction(final int rawShortValue) {
 		double rpmValue = (rawShortValue == 0) ? 0.0 : 1.0 / rawShortValue * 120000000.0;
 		return rpmValue;
 	}
